requirements.txt
aiogram==3.22.0
python-dotenv>=1.0,<2.0


